update db ver fix 环奈 change love panel add time for seal action

+++ Assets/Scrips/Elements/BattleActions/BarrierAction.cs
@@ -34,22 +34,22 @@ namespace Elements
             switch (ActionDetail1)
             {
                 case 1:
-                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.GUARD_ATK, _valueDictionary[eValueNumber.VALUE_3], this, _skill, (float)intReverseTruncate);
+                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.GUARD_ATK, _valueDictionary[eValueNumber.VALUE_3], this, _skill, intReverseTruncate);
                     break;
                 case 2:
-                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.GUARD_MGC, _valueDictionary[eValueNumber.VALUE_3], this, _skill, (float)intReverseTruncate);
+                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.GUARD_MGC, _valueDictionary[eValueNumber.VALUE_3], this, _skill, intReverseTruncate);
                     break;
                 case 3:
-                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.DRAIN_ATK, _valueDictionary[eValueNumber.VALUE_3], this, _skill, (float)intReverseTruncate);
+                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.DRAIN_ATK, _valueDictionary[eValueNumber.VALUE_3], this, _skill, intReverseTruncate);
                     break;
                 case 4:
-                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.DRAIN_MGC, _valueDictionary[eValueNumber.VALUE_3], this, _skill, (float)intReverseTruncate);
+                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.DRAIN_MGC, _valueDictionary[eValueNumber.VALUE_3], this, _skill, intReverseTruncate);
                     break;
                 case 5:
-                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.GUARD_BOTH, _valueDictionary[eValueNumber.VALUE_3], this, _skill, (float)intReverseTruncate);
+                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.GUARD_BOTH, _valueDictionary[eValueNumber.VALUE_3], this, _skill, intReverseTruncate);
                     break;
                 case 6:
-                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.DRAIN_BOTH, _valueDictionary[eValueNumber.VALUE_3], this, _skill, (float)intReverseTruncate);
+                    _target.Owner.SetAbnormalState(_source, UnitCtrl.eAbnormalState.DRAIN_BOTH, _valueDictionary[eValueNumber.VALUE_3], this, _skill, intReverseTruncate);
                     break;
             }
             action?.Invoke("生成可以" + ((eBarrierType)ActionDetail1).GetDescription() + "的护盾，数值" + intReverseTruncate + "持续时间" + _valueDictionary[eValueNumber.VALUE_3]);
